requirements.txt
# DO NOT include torch or torchvision here — install them separately as shown.
opencv-python
Pillow
numpy
scikit-learn
tqdm
fastapi
uvicorn[standard]
streamlit
pytorch-grad-cam
matplotlib
requests
python-multipart
